humanproc_utils/pose_estimation_3d/mediapipe/requirements.txt
# For mp_pose_estimation
json
PyYAML
numpy
matplotlib
mpl_toolkits
mediapipe == 0.10.15
opencv-python == 4.10.0
